behavior-preserving cleanup: test-file restructures, resolution-chain unification, test dedup
Thermo-nuclear review follow-through on the provider-rework branch. Zero
behavior change: production edits are moves/dedup/docs only; the suite
stays green throughout (691 -> 685 tests, each drop justified below).

File restructures (LOC-neutral moves):
- model/mod.rs (2,507 -> 151 lines): the inline `mod tests` moves to
  model/tests/ split by subject (settings_validation, auth, cost,
  anthropic, completions, parsers); shared helpers and the backoff test
  stay in tests/mod.rs. The module-level globs are untouched.
- model/client.rs (1,971 -> client/mod.rs 619 lines): tests move to
  client/tests/ (http_contract, header_policy, s5_wire, cost_attach) with
  shared scripted-server helpers in tests/mod.rs. Kept as children of
  `client` (not a sibling client_tests/) so the tests retain access to
  the module's private items with zero visibility changes.

Production dedup:
- The exact -> dated-family -> provider-default resolution chain was
  implemented three times (ModelCatalog::resolve, user overrides,
  overlay seed-map lookup); all three now share
  ProviderCatalog::resolve_entry.
- The hydrate-and-insert provider loop (baseline merge, overlay merge)
  is now data::merge_entries.
- validate_backend_api_key_env loses its unused _base_url parameter; the
  three identical validate calls in from_effective_settings hoist above
  the match (ArceeApi keeps its base-url-first error precedence).
- `pub mod catalog` is now private and the ModelMetadata/ThinkingLevelMap
  re-exports drop to pub(crate); nothing outside model/ named them
  (server uses only spawn_overlay_refresh).
- The three write-only ModelMetadata fields (source, display_name,
  reasoning) stay for the planned /models endpoint and now say so in
  their field docs; is_explicitly_unsupported likewise (test-consumed,
  documents the user-override null wire-value case).

Test dedup (-6 tests, -248 test lines):
- catalog/tests.rs: the seed matrix guard keeps only unknown-model rows
  (known models and their dated snapshots are guarded exhaustively by
  every_generated_entry_preserves_the_validation_matrix against the same
  independent transcription); validation_resolves_dated_snapshot_families
  (dated resolution already pinned by dated_snapshots_resolve + the
  guards) and unknown_models_keep_the_conservative_provider_default_
  rejection (the trimmed seed guard's unknown rows pin validate() against
  the transcription for every provider) are deleted as subsumed.
- user_override_tests.rs: user_override_thinking_map_relaxes_effort_levels
  deleted; the end-to-end variant in catalog/tests.rs pins the same map
  plus validation, source, and the emitted wire value.
- overlay_tests.rs: the three contained-failure refresh tests merge into
  one case-table test (-2).
- client s5 wire cluster: same_model_history_replays_reasoning_items_on_
  openai_responses dropped (-1; item expansion pinned by the builder
  test, same-origin pass-through by history.rs, full input composition
  by the cross-model wire test). The anthropic same-model golden test
  stays: it is the only full-shape end-to-end pin of the highest-risk
  replay path (thinking blocks + tool_use + tool_result + cache_control).
- mod.rs: the pre-S4 anthropic capability-matrix test trims to its unique
  residue (typed-error contract of a rejected pair; accept/reject
  behavior is guard-covered); the openai-schemas test drops the
  supported-effort loops byte-pinned by the one-builder golden test;
  the three resolve-and-project helpers merge into test_resolved.
- TempHome/write_overlay/EnvGuard scaffolding (three near-copies) unifies
  into catalog/test_support.rs.

Verified: cargo test --workspace --locked 685 passed / 0 failed (2
pre-existing ignored); cargo check --workspace --locked --tests zero
warnings; node --test app.test.js 115/115; nac-catalog-gen --check exit
0; cargo build --workspace --locked --offline clean.
