+++ crates/nac-core/src/model/mod.rs
@@ -20,7 +20,7 @@ mod anthropic;
 mod arcee;
 mod auth_store;
 mod backend;
-pub mod catalog;
+mod catalog;
 mod chat;
 mod chatgpt_codex;
 mod client;
@@ -33,8 +33,8 @@ mod types;
 
 use arcee::{arcee_auth_login, arcee_auth_logout, arcee_auth_status};
 pub use backend::{validate_backend_api_key_env, validate_model_reasoning_effort};
-pub use catalog::{spawn_overlay_refresh, ModelMetadata, ThinkingLevelMap};
-pub(crate) use catalog::{Compat, CompletionsThinkingFormat};
+pub use catalog::spawn_overlay_refresh;
+pub(crate) use catalog::{Compat, CompletionsThinkingFormat, ModelMetadata, ThinkingLevelMap};
 use chatgpt_codex::{codex_auth_login, codex_auth_logout, codex_auth_status};
 pub use client::validate_model_configuration;
 pub(crate) use client::ModelClient;
@@ -148,2360 +148,4 @@ use requests::*;
 use responses::*;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::TEST_ENV_LOCK;
-    use std::ffi::OsString;
-    use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    /// Resolve the catalog thinking map a client's `resolved_model` would
-    /// carry for this backend/model pair (S4 adapter input).
-    fn test_thinking_levels(backend: BackendKind, model: &str) -> ThinkingLevelMap {
-        catalog::resolve(backend, model).thinking_level_map
-    }
-
-    /// Resolve the catalog compat a client's `resolved_model` would carry
-    /// for this backend/model pair (S6 adapter input).
-    fn test_compat(backend: BackendKind, model: &str) -> Compat {
-        catalog::resolve(backend, model).compat
-    }
-
-    /// Resolve the catalog max_tokens a client's `resolved_model` would
-    /// carry for this backend/model pair (S6 adapter input).
-    fn test_max_tokens(backend: BackendKind, model: &str) -> u64 {
-        catalog::resolve(backend, model).max_tokens
-    }
-
-    struct IsolatedModelEnv {
-        original: Vec<(&'static str, Option<OsString>)>,
-        home: PathBuf,
-    }
-
-    impl IsolatedModelEnv {
-        fn new(
-            label: &str,
-            auth_contents: Option<&str>,
-            openai_key: Option<&str>,
-            base_url: Option<&str>,
-        ) -> Self {
-            let unique = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .expect("time went backwards")
-                .as_nanos();
-            let home = std::env::temp_dir()
-                .join(format!("nac-model-{label}-{}-{unique}", std::process::id()));
-            std::fs::create_dir_all(&home).unwrap();
-            if let Some(contents) = auth_contents {
-                write_test_credential(&home.join("auth.json"), contents);
-            }
-
-            let names = [
-                "OPENAI_API_KEY",
-                "OPENAI_BASE_URL",
-                "OPENAI_MODEL",
-                "NAC_HOME",
-            ];
-            let original = names
-                .into_iter()
-                .map(|name| (name, std::env::var_os(name)))
-                .collect();
-            set_env("OPENAI_API_KEY", openai_key);
-            set_env("OPENAI_BASE_URL", base_url);
-            set_env("OPENAI_MODEL", None);
-            unsafe { std::env::set_var("NAC_HOME", &home) };
-
-            Self { original, home }
-        }
-    }
-
-    impl Drop for IsolatedModelEnv {
-        fn drop(&mut self) {
-            for (name, value) in self.original.drain(..) {
-                restore_env(name, value);
-            }
-            let _ = std::fs::remove_dir_all(&self.home);
-        }
-    }
-
-    fn write_test_credential(path: &std::path::Path, contents: impl AsRef<[u8]>) {
-        std::fs::write(path, contents).unwrap();
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(path, std::fs::Permissions::from_mode(0o600)).unwrap();
-        }
-    }
-
-    fn set_env(name: &str, value: Option<&str>) {
-        match value {
-            Some(value) => unsafe { std::env::set_var(name, value) },
-            None => unsafe { std::env::remove_var(name) },
-        }
-    }
-
-    fn stored_arcee_auth(access_token: &str, base_url: &str) -> String {
-        json!({
-            "type": "arcee_device_token",
-            "access_token": access_token,
-            "refresh_token": "refresh-test",
-            "token_type": "bearer",
-            "expires_at_ms": u64::MAX,
-            "base_url": base_url,
-            "organization_id": "org-test",
-            "workspace_name": "workspace-test"
-        })
-        .to_string()
-    }
-
-    fn stored_codex_auth() -> String {
-        json!({
-            "type": "chatgpt-codex",
-            "access": "access-test",
-            "refresh": "refresh-test",
-            "expires_at_ms": u64::MAX,
-            "account_id": "account-test"
-        })
-        .to_string()
-    }
-
-    fn restore_env(name: &str, value: Option<OsString>) {
-        match value {
-            Some(value) => unsafe { std::env::set_var(name, value) },
-            None => unsafe { std::env::remove_var(name) },
-        }
-    }
-
-    fn directory_names(path: &std::path::Path) -> Vec<String> {
-        let mut names = std::fs::read_dir(path)
-            .unwrap()
-            .map(|entry| entry.unwrap().file_name().to_string_lossy().into_owned())
-            .collect::<Vec<_>>();
-        names.sort();
-        names
-    }
-
-    #[test]
-    fn backoff_duration_stays_within_jitter_bounds() {
-        for attempt in 0..10usize {
-            let base_ms = std::cmp::min(200u64.saturating_mul(1 << attempt), 30_000);
-            // Multiplier must span [0.9, 1.1): inclusive bounds with one
-            // ulp of slack for f64 rounding at the top edge.
-            let lower = (base_ms as f64 * 0.9) as u64;
-            let upper = (base_ms as f64 * 1.1) as u64;
-            for _ in 0..200 {
-                let delay = backoff_duration(attempt);
-                assert!(
-                    delay >= Duration::from_millis(lower)
-                        && delay <= Duration::from_millis(upper),
-                    "attempt {attempt} produced {delay:?} outside [{lower}ms, {upper}ms]"
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn api_key_backends_require_explicit_valid_selectors_and_ignore_canonical_vars() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let names = [
-            "OPENAI_API_KEY",
-            "TOGETHER_API_KEY",
-            "ANTHROPIC_API_KEY",
-            "NAC_EXPLICIT_TEST_KEY",
-        ];
-        let original = names.map(|name| (name, std::env::var_os(name)));
-        set_env("OPENAI_API_KEY", Some("openai-selected"));
-        set_env("TOGETHER_API_KEY", Some("together-selected"));
-        set_env("ANTHROPIC_API_KEY", Some("anthropic-selected"));
-        set_env("NAC_EXPLICIT_TEST_KEY", Some("selected-secret"));
-
-        let backends = [
-            BackendKind::OpenAiResponses,
-            BackendKind::TogetherChat,
-            BackendKind::AnthropicMessages,
-            BackendKind::DeepSeekChat,
-            BackendKind::FireworksChat,
-            BackendKind::ArceeApi,
-        ];
-        for backend in backends {
-            let missing = api_key_for_backend(backend, None)
-                .expect_err("canonical variables must not act as implicit selectors");
-            assert!(missing.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(missing
-                .to_string()
-                .contains("requires a nonblank api_key_env"));
-
-            let selected = api_key_for_backend(backend, Some("NAC_EXPLICIT_TEST_KEY"))
-                .expect("explicit selector should be authoritative");
-            assert_eq!(selected, "selected-secret");
-        }
-
-        for (selector, expected) in [
-            ("OPENAI_API_KEY", "openai-selected"),
-            ("TOGETHER_API_KEY", "together-selected"),
-            ("ANTHROPIC_API_KEY", "anthropic-selected"),
-        ] {
-            let selected = api_key_for_backend(BackendKind::OpenAiResponses, Some(selector))
-                .expect("canonical variable names remain valid explicit selectors");
-            assert_eq!(selected, expected);
-        }
-
-        for (name, value) in original {
-            restore_env(name, value);
-        }
-    }
-
-    #[test]
-    fn api_key_selector_validation_and_values_are_typed_configuration_errors() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let names = [
-            "NAC_MISSING_TEST_KEY",
-            "NAC_EMPTY_TEST_KEY",
-            "NAC_SPACE_TEST_KEY",
-        ];
-        let original = names.map(|name| (name, std::env::var_os(name)));
-        set_env("NAC_MISSING_TEST_KEY", None);
-        set_env("NAC_EMPTY_TEST_KEY", Some(""));
-        set_env("NAC_SPACE_TEST_KEY", Some("  \t "));
-
-        for selector in [None, Some(""), Some("   ")] {
-            let error = api_key_for_backend(BackendKind::OpenAiResponses, selector).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains("nonblank api_key_env"));
-        }
-        for selector in ["9INVALID", "HAS-DASH", " HAS_SPACE"] {
-            let error =
-                api_key_for_backend(BackendKind::OpenAiResponses, Some(selector)).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains(selector));
-            assert!(error.to_string().contains("[A-Za-z_][A-Za-z0-9_]*"));
-        }
-        for selector in names {
-            let error =
-                api_key_for_backend(BackendKind::OpenAiResponses, Some(selector)).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains(selector));
-            assert!(!error.to_string().contains("ambient-must-not-win"));
-        }
-
-        for (name, value) in original {
-            restore_env(name, value);
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn non_unicode_selected_api_key_is_a_typed_configuration_error() {
-        use std::os::unix::ffi::OsStringExt;
-
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let name = "NAC_NON_UNICODE_TEST_KEY";
-        let original = std::env::var_os(name);
-        unsafe { std::env::set_var(name, OsString::from_vec(vec![0xff, 0xfe])) };
-
-        let error = api_key_for_backend(BackendKind::OpenAiResponses, Some(name)).unwrap_err();
-        assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-        assert!(error.to_string().contains(name));
-        assert!(error.to_string().contains("non-Unicode"));
-
-        restore_env(name, original);
-    }
-
-    fn effective_settings(
-        backend: BackendKind,
-        base_url: &str,
-        api_key_env: Option<&str>,
-    ) -> EffectiveModelSettings {
-        EffectiveModelSettings::new(
-            backend,
-            "test-model".to_string(),
-            base_url.to_string(),
-            None,
-            api_key_env.map(str::to_string),
-            std::collections::BTreeMap::new(),
-        )
-        .unwrap()
-    }
-
-    #[test]
-    fn arcee_auth_rejects_nonempty_api_key_env_before_credentials() {
-        let expected = "invalid model configuration: api_key_env 'ARCEE_API_KEY' is not supported for backend 'arcee-auth'; managed Arcee auth uses arcee_auth.json";
-        let error = ModelClient::from_effective_settings(
-            EffectiveModelSettings::new(
-                BackendKind::ArceeAuth,
-                "test-model".to_string(),
-                "https://api.arcee.ai".to_string(),
-                None,
-                Some("ARCEE_API_KEY".to_string()),
-                std::collections::BTreeMap::new(),
-            )
-            .unwrap(),
-        )
-        .expect_err("managed Arcee configuration must reject api_key_env");
-        assert_eq!(error.to_string(), expected);
-    }
-
-    #[test]
-    fn backend_kind_parses_and_serializes_explicit_arcee_modes() {
-        for (raw, expected) in [
-            ("arcee-auth", BackendKind::ArceeAuth),
-            ("arcee-api", BackendKind::ArceeApi),
-        ] {
-            let parsed: BackendKind = serde_json::from_str(&format!("\"{raw}\"")).unwrap();
-            assert_eq!(parsed, expected);
-            assert_eq!(parsed.to_string(), raw);
-            assert_eq!(
-                serde_json::to_string(&parsed).unwrap(),
-                format!("\"{raw}\"")
-            );
-        }
-    }
-
-    #[test]
-    fn removed_backend_names_require_settings_repair() {
-        for raw in ["arcee", "auto"] {
-            let error = serde_json::from_str::<BackendKind>(&format!("\"{raw}\""))
-                .unwrap_err()
-                .to_string();
-            assert!(error.contains("unsupported backend"), "{error}");
-            assert!(error.contains("settings repair required"), "{error}");
-        }
-    }
-
-    #[test]
-    fn managed_backends_reject_any_present_api_key_selector() {
-        for (backend, source) in [
-            (BackendKind::ArceeAuth, "arcee_auth.json"),
-            (
-                BackendKind::ChatGptCodexResponses,
-                "stored OAuth from auth.json",
-            ),
-        ] {
-            for selector in ["MANAGED_KEY", "", "   ", " SURROUNDED_KEY "] {
-                let error = validate_backend_api_key_env(
-                    backend,
-                    Some("https://service.example"),
-                    Some(selector),
-                )
-                .expect_err("managed credentials must reject every present api_key_env");
-                assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-                assert!(error.to_string().contains(source));
-                assert!(error.to_string().contains("is not supported"));
-            }
-        }
-    }
-
-    #[test]
-    fn effective_settings_require_explicit_valid_model_tuple() {
-        for (backend, model, base_url, expected) in [
-            (
-                None,
-                Some("model".to_string()),
-                Some("https://example.com".to_string()),
-                "backend",
-            ),
-            (
-                Some(BackendKind::OpenAiResponses),
-                None,
-                Some("https://example.com".to_string()),
-                "model",
-            ),
-            (
-                Some(BackendKind::OpenAiResponses),
-                Some("model".to_string()),
-                None,
-                "base_url",
-            ),
-            (
-                Some(BackendKind::OpenAiResponses),
-                Some("model".to_string()),
-                Some("not a url".to_string()),
-                "base_url",
-            ),
-        ] {
-            let error = EffectiveModelSettings::from_optional(
-                backend,
-                model,
-                base_url,
-                None,
-                None,
-                std::collections::BTreeMap::new(),
-            )
-            .unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains(expected), "{error:#}");
-        }
-    }
-
-    #[test]
-    fn managed_backends_materialize_only_absent_base_urls() {
-        for (backend, expected) in [
-            (
-                BackendKind::ChatGptCodexResponses,
-                CHATGPT_CODEX_CANONICAL_BASE_URL,
-            ),
-            (BackendKind::ArceeAuth, ARCEE_AUTH_CANONICAL_BASE_URL),
-        ] {
-            let materialized = EffectiveModelSettings::from_optional(
-                Some(backend),
-                Some("model".to_string()),
-                None,
-                None,
-                None,
-                std::collections::BTreeMap::new(),
-            )
-            .expect("managed backend should materialize its absent base URL");
-            assert_eq!(materialized.base_url, expected);
-
-            let configured = EffectiveModelSettings::from_optional(
-                Some(backend),
-                Some("model".to_string()),
-                Some(expected.to_string()),
-                None,
-                None,
-                std::collections::BTreeMap::new(),
-            )
-            .expect("matching configured managed base URL should remain accepted");
-            assert_eq!(configured.base_url, expected);
-
-            let error = EffectiveModelSettings::from_optional(
-                Some(backend),
-                Some("model".to_string()),
-                Some("   ".to_string()),
-                None,
-                None,
-                std::collections::BTreeMap::new(),
-            )
-            .expect_err("a present invalid base URL must not be replaced by the default");
-            assert!(error.to_string().contains("must not be blank"), "{error:#}");
-        }
-
-        let error = EffectiveModelSettings::from_optional(
-            Some(BackendKind::OpenAiResponses),
-            Some("model".to_string()),
-            None,
-            None,
-            None,
-            std::collections::BTreeMap::new(),
-        )
-        .expect_err("API-key backends must not acquire a base URL default");
-        assert!(error.to_string().contains("base_url"), "{error:#}");
-    }
-
-    #[test]
-    fn effective_settings_reject_unsupported_reasoning_before_client_or_persistence() {
-        let all = [
-            ReasoningEffort::None,
-            ReasoningEffort::Minimal,
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-            ReasoningEffort::Xhigh,
-        ];
-        let cases: &[(BackendKind, &str, &[ReasoningEffort])] = &[
-            (
-                BackendKind::DeepSeekChat,
-                "model",
-                &[
-                    ReasoningEffort::None,
-                    ReasoningEffort::High,
-                    ReasoningEffort::Xhigh,
-                ],
-            ),
-            (
-                BackendKind::FireworksChat,
-                "model",
-                &[
-                    ReasoningEffort::None,
-                    ReasoningEffort::Low,
-                    ReasoningEffort::Medium,
-                    ReasoningEffort::High,
-                ],
-            ),
-            (
-                BackendKind::TogetherChat,
-                "model",
-                &[
-                    ReasoningEffort::None,
-                    ReasoningEffort::Low,
-                    ReasoningEffort::Medium,
-                    ReasoningEffort::High,
-                ],
-            ),
-            (BackendKind::OpenAiResponses, "model", &all),
-            (BackendKind::ChatGptCodexResponses, "model", &all),
-            (
-                BackendKind::AnthropicMessages,
-                "claude-opus-4-6",
-                &[
-                    ReasoningEffort::None,
-                    ReasoningEffort::Low,
-                    ReasoningEffort::Medium,
-                    ReasoningEffort::High,
-                    ReasoningEffort::Xhigh,
-                ],
-            ),
-            (BackendKind::ArceeAuth, "model", &[]),
-            (BackendKind::ArceeApi, "model", &[]),
-        ];
-
-        for (backend, model, supported) in cases {
-            EffectiveModelSettings::new(
-                *backend,
-                (*model).into(),
-                "https://example.com/v1".into(),
-                None,
-                None,
-                std::collections::BTreeMap::new(),
-            )
-            .expect("absent effort must be valid for every backend");
-            for effort in all {
-                let result = EffectiveModelSettings::new(
-                    *backend,
-                    (*model).into(),
-                    "https://example.com/v1".into(),
-                    Some(effort),
-                    None,
-                    std::collections::BTreeMap::new(),
-                );
-                if supported.contains(&effort) {
-                    result.unwrap_or_else(|error| {
-                        panic!("{backend} rejected {}: {error:#}", effort.as_str())
-                    });
-                } else {
-                    let error = result
-                        .expect_err("unsupported effort must fail effective settings validation");
-                    assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-                    assert!(error.to_string().contains(effort.as_str()), "{error:#}");
-                    assert!(error.to_string().contains(backend.as_str()), "{error:#}");
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn anthropic_reasoning_capability_matrix_is_model_dependent_and_conservative() {
-        let all = [
-            ReasoningEffort::None,
-            ReasoningEffort::Minimal,
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-            ReasoningEffort::Xhigh,
-        ];
-        let standard = [
-            ReasoningEffort::None,
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-        ];
-        let with_max = [
-            ReasoningEffort::None,
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-            ReasoningEffort::Xhigh,
-        ];
-        let none_only = [ReasoningEffort::None];
-        let cases: &[(&str, &[ReasoningEffort])] = &[
-            ("claude-opus-4-6", &with_max),
-            ("claude-opus-4-6-20260205", &with_max),
-            ("claude-sonnet-4-6", &standard),
-            ("claude-sonnet-4-6-20260217", &standard),
-            ("claude-opus-4-5", &none_only),
-            ("claude-sonnet-4-5", &none_only),
-            ("claude-opus-4-6-latest", &none_only),
-            ("claude-always-on-future", &none_only),
-        ];
-
-        for (model, supported) in cases {
-            validate_model_reasoning_effort(BackendKind::AnthropicMessages, model, None)
-                .expect("absent effort must never select Anthropic thinking controls");
-            for effort in all {
-                let result = validate_model_reasoning_effort(
-                    BackendKind::AnthropicMessages,
-                    model,
-                    Some(effort),
-                );
-                if supported.contains(&effort) {
-                    result.unwrap_or_else(|error| {
-                        panic!(
-                            "Anthropic model {model} rejected {}: {error:#}",
-                            effort.as_str()
-                        )
-                    });
-                } else {
-                    let error = result.expect_err("unsupported model/effort pair must fail");
-                    assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-                    assert!(error.to_string().contains(model), "{error:#}");
-                    assert!(error.to_string().contains(effort.as_str()), "{error:#}");
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn validation_error_messages_are_preserved_verbatim() {
-        // S4 derives the "supported values" list from the model's catalog
-        // map; the user-facing error strings are byte-identical to the
-        // pre-S4 matrix errors.
-        let cases: &[(BackendKind, &str, ReasoningEffort, &str)] = &[
-            (
-                BackendKind::DeepSeekChat,
-                "model",
-                ReasoningEffort::Low,
-                "invalid model configuration: reasoning effort 'low' is not supported by backend 'deepseek-chat'; supported values: none, high, or xhigh",
-            ),
-            (
-                BackendKind::FireworksChat,
-                "model",
-                ReasoningEffort::Xhigh,
-                "invalid model configuration: reasoning effort 'xhigh' is not supported by backend 'fireworks-chat'; supported values: none, low, medium, or high",
-            ),
-            (
-                BackendKind::TogetherChat,
-                "model",
-                ReasoningEffort::Minimal,
-                "invalid model configuration: reasoning effort 'minimal' is not supported by backend 'together-chat'; supported values: none, low, medium, or high",
-            ),
-            (
-                BackendKind::AnthropicMessages,
-                "claude-opus-4-6",
-                ReasoningEffort::Minimal,
-                "invalid model configuration: reasoning effort 'minimal' is not supported by backend 'anthropic-messages' for Anthropic model 'claude-opus-4-6'; supported values: none, low, medium, high, or xhigh",
-            ),
-            (
-                BackendKind::AnthropicMessages,
-                "claude-sonnet-4-6",
-                ReasoningEffort::Xhigh,
-                "invalid model configuration: reasoning effort 'xhigh' is not supported by backend 'anthropic-messages' for Anthropic model 'claude-sonnet-4-6'; supported values: none, low, medium, or high",
-            ),
-            (
-                BackendKind::AnthropicMessages,
-                "claude-opus-4-5",
-                ReasoningEffort::High,
-                "invalid model configuration: reasoning effort 'high' is not supported by backend 'anthropic-messages' for Anthropic model 'claude-opus-4-5'; supported values: none only",
-            ),
-            (
-                BackendKind::AnthropicMessages,
-                "claude-opus-4-6-latest",
-                ReasoningEffort::High,
-                "invalid model configuration: reasoning effort 'high' is not supported by backend 'anthropic-messages' for Anthropic model 'claude-opus-4-6-latest'; supported values: none only",
-            ),
-            (
-                BackendKind::ArceeAuth,
-                "model",
-                ReasoningEffort::None,
-                "invalid model configuration: reasoning effort 'none' is not supported by backend 'arcee-auth'; supported values: no explicit effort levels",
-            ),
-        ];
-        for (backend, model, effort, expected) in cases {
-            let error = validate_model_reasoning_effort(*backend, model, Some(*effort))
-                .expect_err("case must be rejected");
-            assert_eq!(error.to_string(), *expected, "{backend} {model}");
-        }
-    }
-
-    #[test]
-    fn adapters_translate_effort_through_the_catalog_map() {
-        // The wire value comes from the passed catalog map, not from adapter
-        // code: a custom map with non-standard wire tiers flows through
-        // every completions/responses/anthropic builder. (requests.rs and
-        // anthropic.rs carry no "max" literal; DeepSeek's and Anthropic's
-        // top tiers are data.)
-        let custom = ThinkingLevelMap(std::collections::BTreeMap::from([
-            (ReasoningEffort::None, Some("none".to_string())),
-            (ReasoningEffort::High, Some("tier-three".to_string())),
-            (ReasoningEffort::Xhigh, Some("tier-four".to_string())),
-        ]));
-        let messages = [Message::User {
-            content: "hi".to_string(),
-        }];
-
-        let deepseek = completions_chat_request(
-            "m",
-            Some(ReasoningEffort::Xhigh),
-            &messages,
-            &[],
-            &custom,
-            &test_compat(BackendKind::DeepSeekChat, "m"),
-        );
-        assert_eq!(deepseek["thinking"], json!({"type": "enabled"}));
-        assert_eq!(deepseek["reasoning_effort"], "tier-four");
-
-        let fireworks = completions_chat_request(
-            "m",
-            Some(ReasoningEffort::High),
-            &messages,
-            &[],
-            &custom,
-            &test_compat(BackendKind::FireworksChat, "m"),
-        );
-        assert_eq!(fireworks["reasoning_effort"], "tier-three");
-        assert_eq!(fireworks["reasoning_history"], "preserved");
-
-        let together = completions_chat_request(
-            "m",
-            Some(ReasoningEffort::High),
-            &messages,
-            &[],
-            &custom,
-            &test_compat(BackendKind::TogetherChat, "m"),
-        );
-        assert_eq!(together["reasoning"], json!({"enabled": true}));
-        assert_eq!(together["reasoning_effort"], "tier-three");
-
-        let openai =
-            openai_responses_request("m", Some(ReasoningEffort::Xhigh), &messages, &[], &custom);
-        assert_eq!(openai["reasoning"]["effort"], "tier-four");
-
-        // claude-opus-4-6 supports xhigh in the baseline catalog, so the
-        // builder's defense-in-depth validation passes; the custom map then
-        // drives the emitted wire tier.
-        let anthropic = anthropic_messages_request(
-            "claude-opus-4-6",
-            Some(ReasoningEffort::Xhigh),
-            &messages,
-            &[],
-            None,
-            &custom,
-            test_max_tokens(BackendKind::AnthropicMessages, "claude-opus-4-6"),
-        )
-        .unwrap();
-        assert_eq!(anthropic["thinking"], json!({"type": "adaptive"}));
-        assert_eq!(anthropic["output_config"]["effort"], "tier-four");
-    }
-
-    #[test]
-    fn stored_codex_auth_config_and_store_failures_remain_distinct() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let settings = effective_settings(
-            BackendKind::ChatGptCodexResponses,
-            "https://chatgpt.com/backend-api",
-            None,
-        );
-
-        for (label, contents, expected) in [
-            ("missing-codex-auth", None, "not configured"),
-            (
-                "malformed-codex-auth",
-                Some("{not-json}"),
-                "failed to parse",
-            ),
-            (
-                "wrong-provider-codex-auth",
-                Some(r#"{"type":"other"}"#),
-                "provider type",
-            ),
-            (
-                "blank-codex-auth",
-                Some(
-                    r#"{"type":"chatgpt-codex","access":"secret-not-for-errors","refresh":" ","expires_at_ms":1,"account_id":"account"}"#,
-                ),
-                "nonblank field 'refresh'",
-            ),
-        ] {
-            let _env = IsolatedModelEnv::new(label, contents, None, None);
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains(expected), "{error:#}");
-            assert!(!error.to_string().contains("secret-not-for-errors"));
-        }
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let env = IsolatedModelEnv::new(
-                "codex-auth-unsafe-permissions",
-                Some(&stored_codex_auth()),
-                None,
-                None,
-            );
-            std::fs::set_permissions(
-                env.home.join("auth.json"),
-                std::fs::Permissions::from_mode(0o644),
-            )
-            .unwrap();
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains("unsafe permissions 0644"));
-            assert!(error.to_string().contains("mode to 0600"));
-            assert!(!format!("{error:#}").contains("access-test"));
-        }
-
-        {
-            let env = IsolatedModelEnv::new("codex-auth-path-io", None, None, None);
-            std::fs::create_dir(env.home.join("auth.json")).unwrap();
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_none());
-            assert_eq!(error.to_string(), "failed to load stored Codex credentials");
-        }
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::symlink;
-            let env =
-                IsolatedModelEnv::new("codex-lock-symlink", Some(&stored_codex_auth()), None, None);
-            let target = env.home.join("lock-target");
-            std::fs::write(&target, "unchanged").unwrap();
-            symlink(&target, env.home.join("auth.auth.json.lock")).unwrap();
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_none());
-            assert_eq!(error.to_string(), "failed to load stored Codex credentials");
-            assert!(format!("{error:#}").contains("symlink auth lock"));
-            assert_eq!(std::fs::read_to_string(target).unwrap(), "unchanged");
-        }
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::symlink;
-            let env = IsolatedModelEnv::new("codex-auth-symlink", None, None, None);
-            let target = env.home.join("target.json");
-            std::fs::write(&target, stored_codex_auth()).unwrap();
-            symlink(&target, env.home.join("auth.json")).unwrap();
-            let error = ModelClient::from_effective_settings(settings).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_none());
-            assert_eq!(error.to_string(), "failed to load stored Codex credentials");
-            assert!(format!("{error:#}").contains("symlink credential path"));
-        }
-    }
-
-    #[test]
-    fn invalid_codex_endpoint_fails_before_credentials_or_connection() {
-        use std::net::TcpListener;
-
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let listener = TcpListener::bind(("127.0.0.1", 0)).unwrap();
-        listener.set_nonblocking(true).unwrap();
-        let endpoint = format!("http://{}/backend-api", listener.local_addr().unwrap());
-        let _env = IsolatedModelEnv::new("codex-no-connection", None, None, None);
-
-        let error = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ChatGptCodexResponses,
-            &endpoint,
-            None,
-        ))
-        .unwrap_err();
-
-        assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-        assert!(error.to_string().contains("requires HTTPS"));
-        std::thread::sleep(std::time::Duration::from_millis(50));
-        assert!(listener.accept().is_err(), "invalid endpoint was contacted");
-    }
-
-    #[test]
-    fn stored_arcee_auth_config_and_store_failures_remain_distinct() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let settings = effective_settings(BackendKind::ArceeAuth, "https://api.arcee.ai", None);
-
-        {
-            let _env = IsolatedModelEnv::new("missing-stored-auth", None, None, None);
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains("Arcee auth is not configured"));
-        }
-        {
-            let env = IsolatedModelEnv::new("malformed-stored-auth", None, None, None);
-            write_test_credential(&env.home.join("arcee_auth.json"), "{not-json}");
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error
-                .to_string()
-                .contains("failed to parse stored Arcee auth"));
-        }
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let env = IsolatedModelEnv::new("arcee-auth-unsafe-permissions", None, None, None);
-            write_test_credential(
-                &env.home.join("arcee_auth.json"),
-                stored_arcee_auth("secret-not-for-errors", "https://api.arcee.ai"),
-            );
-            std::fs::set_permissions(
-                env.home.join("arcee_auth.json"),
-                std::fs::Permissions::from_mode(0o660),
-            )
-            .unwrap();
-            let error = ModelClient::from_effective_settings(settings.clone()).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_some());
-            assert!(error.to_string().contains("unsafe permissions 0660"));
-            assert!(error.to_string().contains("mode to 0600"));
-            assert!(!format!("{error:#}").contains("secret-not-for-errors"));
-        }
-        {
-            let env = IsolatedModelEnv::new("stored-auth-path-io", None, None, None);
-            std::fs::create_dir(env.home.join("arcee_auth.json")).unwrap();
-            let error = ModelClient::from_effective_settings(settings).unwrap_err();
-            assert!(error.downcast_ref::<ModelConfigurationError>().is_none());
-            assert_eq!(error.to_string(), "failed to load stored Arcee credentials");
-        }
-    }
-
-    #[test]
-    fn explicit_arcee_backend_binds_stored_key_to_its_origin() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let auth = stored_arcee_auth("rcai-test", "https://stored.arcee.ai");
-        let env = IsolatedModelEnv::new("explicit-arcee", None, None, None);
-        write_test_credential(&env.home.join("arcee_auth.json"), &auth);
-
-        let requested_base = "https://stored.arcee.ai:443/api/v1/";
-        let client = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ArceeAuth,
-            requested_base,
-            None,
-        ))
-        .expect("the stored credential should work on the same approved origin");
-        assert_eq!(client.base_url(), requested_base);
-
-        let mismatch = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ArceeAuth,
-            "https://api.internal.arcee.ai/api",
-            None,
-        ))
-        .unwrap_err();
-        assert!(mismatch
-            .to_string()
-            .contains("does not match the stored credential origin"));
-    }
-
-    #[tokio::test]
-    async fn existing_arcee_client_rejects_credentials_rotated_to_another_origin() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let initial_token = "initial-token-must-not-leak";
-        let rotated_token = "rotated-token-must-not-leak";
-        let env = IsolatedModelEnv::new("rotated-arcee-origin", None, None, None);
-        let auth_path = env.home.join("arcee_auth.json");
-        write_test_credential(
-            &auth_path,
-            stored_arcee_auth(initial_token, "https://api.arcee.ai"),
-        );
-        let client = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ArceeAuth,
-            "https://api.arcee.ai/api/v1",
-            None,
-        ))
-        .expect("initial credential origin should match the session");
-
-        write_test_credential(
-            &auth_path,
-            stored_arcee_auth(rotated_token, "https://tenant.arcee.ai"),
-        );
-
-        let fresh_error = client
-            .send_turn(Vec::new(), Vec::new())
-            .await
-            .expect_err("a fresh reload must reject a credential from another origin");
-        let forced_error = arcee::force_refresh_access_token(
-            &arcee::no_redirect_client().unwrap(),
-            client.base_url(),
-            initial_token,
-        )
-        .await
-        .expect_err("a forced refresh must reject a credential from another origin");
-
-        for error in [fresh_error, forced_error] {
-            let diagnostic = format!("{error:#}");
-            assert!(
-                diagnostic.contains("does not match the stored credential origin"),
-                "unexpected origin mismatch: {diagnostic}"
-            );
-            assert!(!diagnostic.contains(initial_token));
-            assert!(!diagnostic.contains(rotated_token));
-        }
-    }
-
-    #[test]
-    fn both_arcee_modes_validate_endpoints_and_sensitive_headers() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let auth = stored_arcee_auth("stored-arcee-secret", "https://api.arcee.ai");
-        let env = IsolatedModelEnv::new("canonical-modes", None, None, None);
-        write_test_credential(&env.home.join("arcee_auth.json"), &auth);
-        let selector = "NAC_ARCEE_CANONICAL_TEST_KEY";
-        let original = std::env::var_os(selector);
-        set_env(selector, Some("api-arcee-secret"));
-
-        for (backend, api_key_env) in [
-            (BackendKind::ArceeAuth, None),
-            (BackendKind::ArceeApi, Some(selector)),
-        ] {
-            let endpoint_error = ModelClient::from_effective_settings(effective_settings(
-                backend,
-                "https://not-arcee.example/v1",
-                api_key_env,
-            ))
-            .unwrap_err();
-            assert!(endpoint_error
-                .to_string()
-                .contains("not an approved Arcee origin"));
-
-            let mut settings = effective_settings(backend, "https://api.arcee.ai", api_key_env);
-            settings
-                .extra_headers
-                .insert("Authorization".to_string(), "must-not-override".to_string());
-            let header_error = ModelClient::from_effective_settings(settings).unwrap_err();
-            assert!(header_error.to_string().contains("Authorization"));
-        }
-        restore_env(selector, original);
-    }
-
-    #[test]
-    fn arcee_api_uses_only_the_selected_variable() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let selector = "NAC_ARCEE_API_TEST_KEY";
-        let original = std::env::var_os(selector);
-        set_env(selector, Some("arcee-api-selected-secret"));
-        let env = IsolatedModelEnv::new("arcee-api", None, None, None);
-        write_test_credential(&env.home.join("arcee_auth.json"), "{not-json}");
-
-        let client = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ArceeApi,
-            "https://api.arcee.ai/api/v1",
-            Some(selector),
-        ))
-        .expect("arcee-api should use only the selected variable");
-        assert_eq!(client.backend(), BackendKind::ArceeApi);
-        restore_env(selector, original);
-    }
-
-    #[test]
-    fn tampered_stored_arcee_url_is_rejected_before_client_creation() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let tampered = stored_arcee_auth("rcai-never-use", "https://attacker.example/steal");
-        let env = IsolatedModelEnv::new("tampered-stored-url", None, None, None);
-        write_test_credential(&env.home.join("arcee_auth.json"), tampered);
-
-        let error = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ArceeAuth,
-            "https://api.arcee.ai",
-            None,
-        ))
-        .unwrap_err();
-        assert!(error.to_string().contains("invalid base_url"));
-    }
-
-    #[test]
-    fn arcee_and_codex_auth_coexist_and_logout_independently() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let codex = stored_codex_auth();
-        let arcee = stored_arcee_auth("rcai-test", "https://api.arcee.ai");
-        let env = IsolatedModelEnv::new("coexist", Some(&codex), None, None);
-        write_test_credential(&env.home.join("arcee_auth.json"), &arcee);
-
-        let loaded = arcee::read_stored_auth().unwrap();
-        assert_eq!(loaded.access_token, "rcai-test");
-        codex_auth_status().unwrap();
-
-        arcee_auth_logout().unwrap();
-        assert!(!env.home.join("arcee_auth.json").exists());
-        assert_eq!(
-            std::fs::read_to_string(env.home.join("auth.json")).unwrap(),
-            codex
-        );
-
-        write_test_credential(&env.home.join("arcee_auth.json"), &arcee);
-        codex_auth_logout().unwrap();
-        assert!(!env.home.join("auth.json").exists());
-        assert_eq!(
-            std::fs::read_to_string(env.home.join("arcee_auth.json")).unwrap(),
-            arcee
-        );
-    }
-
-    #[test]
-    fn legacy_shaped_auth_json_is_ignored_and_unchanged_by_arcee_status_and_client_creation() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let legacy = stored_arcee_auth("rcai-legacy", "https://api.arcee.ai");
-        let env = IsolatedModelEnv::new("legacy-ignored", Some(&legacy), None, None);
-        let auth_path = env.home.join("auth.json");
-        let canonical_path = env.home.join("arcee_auth.json");
-        let before = std::fs::read(&auth_path).unwrap();
-
-        arcee_auth_status().unwrap();
-        assert_eq!(std::fs::read(&auth_path).unwrap(), before);
-        assert!(!canonical_path.exists());
-        assert_eq!(directory_names(&env.home), ["auth.json"]);
-
-        let error = ModelClient::from_effective_settings(effective_settings(
-            BackendKind::ArceeAuth,
-            "https://api.arcee.ai",
-            None,
-        ))
-        .expect_err("legacy auth.json must not authenticate arcee-auth");
-        assert!(error.to_string().contains("Arcee auth is not configured"));
-        assert_eq!(std::fs::read(&auth_path).unwrap(), before);
-        assert!(!canonical_path.exists());
-        assert_eq!(directory_names(&env.home), ["auth.json"]);
-    }
-
-    #[test]
-    fn codex_status_and_logout_ignore_legacy_shaped_arcee_auth_json() {
-        let _guard = TEST_ENV_LOCK.lock().unwrap();
-        let legacy = stored_arcee_auth("rcai-legacy", "https://api.arcee.ai");
-        let env = IsolatedModelEnv::new("codex-foreign-arcee", Some(&legacy), None, None);
-        let auth_path = env.home.join("auth.json");
-        let canonical_path = env.home.join("arcee_auth.json");
-        let before = std::fs::read(&auth_path).unwrap();
-
-        codex_auth_status().unwrap();
-        assert_eq!(directory_names(&env.home), ["auth.json"]);
-        codex_auth_logout().unwrap();
-
-        assert_eq!(std::fs::read(&auth_path).unwrap(), before);
-        assert!(!canonical_path.exists());
-    }
-
-    #[test]
-    fn token_usage_validation_and_accumulation_are_overflow_safe() {
-        let valid = TokenUsage {
-            input_tokens: 10,
-            output_tokens: 5,
-            cache_read_tokens: 2,
-            cache_write_tokens: 3,
-            reasoning_tokens: 4,
-            orchestrator_context_tokens: 20,
-            cost: TokenCostMicros::default(),
-        };
-        assert_eq!(valid.valid_provider_context(), Some(20));
-
-        let mut inconsistent = valid.clone();
-        inconsistent.orchestrator_context_tokens = 19;
-        assert_eq!(inconsistent.valid_provider_context(), None);
-
-        let mut maximum_total = valid.clone();
-        maximum_total.orchestrator_context_tokens = crate::MAX_SUPPORTED_TOKEN_COUNT;
-        assert_eq!(
-            maximum_total.valid_provider_context(),
-            Some(crate::MAX_SUPPORTED_TOKEN_COUNT)
-        );
-
-        let mut oversized_total = valid.clone();
-        oversized_total.orchestrator_context_tokens = crate::MAX_SUPPORTED_TOKEN_COUNT + 1;
-        assert_eq!(oversized_total.valid_provider_context(), None);
-
-        let hostile = TokenUsage {
-            input_tokens: u64::MAX,
-            output_tokens: u64::MAX,
-            cache_read_tokens: u64::MAX,
-            cache_write_tokens: u64::MAX,
-            reasoning_tokens: u64::MAX,
-            orchestrator_context_tokens: u64::MAX,
-            cost: TokenCostMicros {
-                input: u64::MAX,
-                output: u64::MAX,
-                cache_read: u64::MAX,
-                cache_write: u64::MAX,
-                total: u64::MAX,
-            },
-        };
-        assert_eq!(hostile.valid_provider_context(), None);
-        let mut accumulated = hostile.clone();
-        accumulated.add_cost_saturating(&hostile);
-        accumulated += hostile;
-        assert_eq!(accumulated.input_tokens, u64::MAX);
-        assert_eq!(accumulated.output_tokens, u64::MAX);
-        assert_eq!(accumulated.orchestrator_context_tokens, u64::MAX);
-        assert_eq!(accumulated.cost.input, u64::MAX);
-        assert_eq!(accumulated.cost.cache_write, u64::MAX);
-        assert_eq!(accumulated.cost.total, u64::MAX);
-    }
-
-    #[test]
-    fn calculate_cost_bills_each_bucket_at_its_catalog_rate() {
-        // claude-opus-4-6 catalog rates ($/1M): 5 / 25 / 0.5 / 6.25.
-        let rates = catalog::ModelCostRates {
-            input: 5.0,
-            output: 25.0,
-            cache_read: 0.5,
-            cache_write: 6.25,
-        };
-        let usage = TokenUsage {
-            input_tokens: 100,
-            output_tokens: 50,
-            cache_read_tokens: 200,
-            cache_write_tokens: 32,
-            reasoning_tokens: 0,
-            orchestrator_context_tokens: 382,
-            cost: TokenCostMicros::default(),
-        };
-
-        // The identity: cost_micros = tokens x rate_per_mtok, exactly.
-        let cost = calculate_cost(&rates, None, &usage);
-        assert_eq!(cost.input, 500);
-        assert_eq!(cost.output, 1_250);
-        assert_eq!(cost.cache_read, 100);
-        assert_eq!(cost.cache_write, 200);
-        assert_eq!(cost.total, 2_050);
-    }
-
-    #[test]
-    fn calculate_cost_rounds_half_up_once_at_the_micro_conversion() {
-        let rates = catalog::ModelCostRates {
-            input: 0.5,
-            output: 1.5,
-            cache_read: 0.49,
-            cache_write: 0.25,
-        };
-        let usage = TokenUsage {
-            input_tokens: 1,
-            output_tokens: 3,
-            cache_read_tokens: 1,
-            cache_write_tokens: 2,
-            reasoning_tokens: 0,
-            orchestrator_context_tokens: 0,
-            cost: TokenCostMicros::default(),
-        };
-
-        let cost = calculate_cost(&rates, None, &usage);
-        assert_eq!(cost.input, 1, "0.5 micros rounds half-up to 1");
-        assert_eq!(cost.output, 5, "4.5 micros rounds half-up to 5");
-        assert_eq!(cost.cache_read, 0, "0.49 micros rounds down to 0");
-        assert_eq!(cost.cache_write, 1, "0.5 micros rounds half-up to 1");
-        assert_eq!(cost.total, 7);
-    }
-
-    #[test]
-    fn calculate_cost_saturates_and_treats_hostile_rates_as_zero() {
-        let usage = TokenUsage {
-            input_tokens: u64::MAX,
-            output_tokens: 10,
-            cache_read_tokens: 10,
-            cache_write_tokens: 10,
-            reasoning_tokens: 0,
-            orchestrator_context_tokens: 0,
-            cost: TokenCostMicros::default(),
-        };
-
-        // Unknown pricing (all-zero rates) bills zero, never an error.
-        assert_eq!(
-            calculate_cost(&catalog::ModelCostRates::default(), None, &usage),
-            TokenCostMicros::default()
-        );
-
-        // Negative, NaN and infinite rates bill zero; only finite positive
-        // rates bill.
-        let hostile = catalog::ModelCostRates {
-            input: -1.0,
-            output: f64::NAN,
-            cache_read: f64::INFINITY,
-            cache_write: 1.0,
-        };
-        let cost = calculate_cost(&hostile, None, &usage);
-        assert_eq!(cost.input, 0);
-        assert_eq!(cost.output, 0);
-        assert_eq!(cost.cache_read, 0);
-        assert_eq!(cost.cache_write, 10);
-
-        // Enormous billable usage saturates at u64::MAX instead of wrapping,
-        // and the stored total saturates rather than overflowing.
-        let enormous = catalog::ModelCostRates {
-            input: 1_000_000_000.0,
-            output: 1_000_000_000.0,
-            cache_read: 1_000_000_000.0,
-            cache_write: 1_000_000_000.0,
-        };
-        let cost = calculate_cost(&enormous, None, &usage);
-        assert_eq!(cost.input, u64::MAX);
-        assert_eq!(cost.output, 10_000_000_000);
-        assert_eq!(cost.total, u64::MAX);
-    }
-
-    #[test]
-    fn calculate_cost_bills_1h_cache_writes_at_the_1h_rate() {
-        let rates = catalog::ModelCostRates {
-            input: 5.0,
-            output: 25.0,
-            cache_read: 0.5,
-            cache_write: 6.25,
-        };
-        let usage = TokenUsage {
-            input_tokens: 0,
-            output_tokens: 0,
-            cache_read_tokens: 0,
-            cache_write_tokens: 100,
-            reasoning_tokens: 0,
-            orchestrator_context_tokens: 0,
-            cost: TokenCostMicros::default(),
-        };
-
-        // 5-minute writes bill at the standard cache_write rate.
-        assert_eq!(calculate_cost(&rates, None, &usage).cache_write, 625);
-        // 1-hour writes bill at the 1h rate (here the 2x-input default).
-        assert_eq!(
-            calculate_cost(&rates, Some(10.0), &usage).cache_write,
-            1_000
-        );
-        // An explicit zero 1h rate bills zero (not the standard rate).
-        assert_eq!(calculate_cost(&rates, Some(0.0), &usage).cache_write, 0);
-    }
-
-    #[test]
-    fn old_token_usage_rows_deserialize_with_zero_cost() {
-        // The pre-S3 persisted shape: no `cost` field anywhere.
-        let old_usage = json!({
-            "input_tokens": 100,
-            "output_tokens": 50,
-            "cache_read_tokens": 200,
-            "cache_write_tokens": 30,
-            "reasoning_tokens": 10,
-            "total_tokens": 380
-        });
-        let usage: TokenUsage = serde_json::from_value(old_usage).unwrap();
-        assert_eq!(usage.input_tokens, 100);
-        assert_eq!(usage.orchestrator_context_tokens, 380);
-        assert_eq!(usage.cost, TokenCostMicros::default());
-
-        // A stored token_usages_json row (Vec<Option<TokenUsage>>) from before
-        // S3 parses with zero cost.
-        let old_row = json!([
-            {
-                "input_tokens": 10,
-                "output_tokens": 20,
-                "cache_read_tokens": 3,
-                "cache_write_tokens": 4,
-                "total_tokens": 37
-            },
-            null
-        ]);
-        let row: Vec<Option<TokenUsage>> = serde_json::from_value(old_row).unwrap();
-        assert_eq!(row.len(), 2);
-        assert_eq!(row[0].as_ref().unwrap().cost, TokenCostMicros::default());
-        assert!(row[1].is_none());
-
-        // Partial cost records fill missing buckets with zero, and the event
-        // JSON shape stays additive (cost rides inside TokenUsage).
-        let partial = json!({
-            "input_tokens": 1,
-            "output_tokens": 2,
-            "cache_read_tokens": 0,
-            "cache_write_tokens": 0,
-            "total_tokens": 3,
-            "cost": {"input": 5, "total": 5}
-        });
-        let usage: TokenUsage = serde_json::from_value(partial).unwrap();
-        assert_eq!(usage.cost.input, 5);
-        assert_eq!(usage.cost.output, 0);
-        assert_eq!(usage.cost.total, 5);
-    }
-
-    #[test]
-    fn anthropic_request_omits_none_and_maps_supported_efforts_exactly() {
-        let messages = [Message::User {
-            content: "read a file".to_string(),
-        }];
-        let unknown_levels =
-            test_thinking_levels(BackendKind::AnthropicMessages, "claude-always-on-future");
-        let unknown_max_tokens =
-            test_max_tokens(BackendKind::AnthropicMessages, "claude-always-on-future");
-        for effort in [None, Some(ReasoningEffort::None)] {
-            let request = anthropic_messages_request(
-                "claude-always-on-future",
-                effort,
-                &messages,
-                &[],
-                None,
-                &unknown_levels,
-                unknown_max_tokens,
-            )
-            .unwrap();
-            // S6: max_tokens is the resolved catalog value — the conservative
-            // 16_384 fallback for a model with no catalog entry (previously
-            // the hardcoded 128_000 for every Anthropic model).
-            assert_eq!(request["max_tokens"], 16_384);
-            assert!(request.get("thinking").is_none());
-            assert!(request.get("output_config").is_none());
-            assert!(!request.to_string().contains("disabled"));
-        }
-
-        let opus_levels = test_thinking_levels(BackendKind::AnthropicMessages, "claude-opus-4-6");
-        for (effort, wire_effort) in [
-            (ReasoningEffort::Low, "low"),
-            (ReasoningEffort::Medium, "medium"),
-            (ReasoningEffort::High, "high"),
-            (ReasoningEffort::Xhigh, "max"),
-        ] {
-            let request = anthropic_messages_request(
-                "claude-opus-4-6",
-                Some(effort),
-                &messages,
-                &[],
-                None,
-                &opus_levels,
-                test_max_tokens(BackendKind::AnthropicMessages, "claude-opus-4-6"),
-            )
-            .unwrap();
-            assert_eq!(request["thinking"], json!({"type": "adaptive"}));
-            assert_eq!(request["output_config"], json!({"effort": wire_effort}));
-        }
-    }
-
-    #[test]
-    fn anthropic_request_with_1h_ttl_sets_ttl_on_all_breakpoints() {
-        let request = anthropic_messages_request(
-            "claude-sonnet-4-6",
-            None,
-            &[
-                Message::System {
-                    content: "system".to_string(),
-                },
-                Message::User {
-                    content: "hello".to_string(),
-                },
-            ],
-            &[ToolDefinition {
-                def_type: "function".to_string(),
-                function: crate::types::FunctionDef {
-                    name: "read".to_string(),
-                    description: "Read".to_string(),
-                    parameters: json!({"type": "object"}),
-                },
-            }],
-            Some("1h"),
-            &test_thinking_levels(BackendKind::AnthropicMessages, "claude-sonnet-4-6"),
-            test_max_tokens(BackendKind::AnthropicMessages, "claude-sonnet-4-6"),
-        )
-        .unwrap();
-
-        // System breakpoint has 1h TTL.
-        assert_eq!(request["system"][0]["cache_control"]["type"], "ephemeral");
-        assert_eq!(request["system"][0]["cache_control"]["ttl"], "1h");
-        // Tool breakpoint has 1h TTL.
-        assert_eq!(request["tools"][0]["cache_control"]["type"], "ephemeral");
-        assert_eq!(request["tools"][0]["cache_control"]["ttl"], "1h");
-        // Last message breakpoint has 1h TTL.
-        assert_eq!(
-            request["messages"][0]["content"][0]["cache_control"]["ttl"],
-            "1h"
-        );
-    }
-
-    #[test]
-    fn anthropic_request_with_no_messages_skips_message_breakpoint() {
-        let request = anthropic_messages_request(
-            "claude-sonnet-4-6",
-            None,
-            &[Message::System {
-                content: "system only".to_string(),
-            }],
-            &[],
-            None,
-            &test_thinking_levels(BackendKind::AnthropicMessages, "claude-sonnet-4-6"),
-            test_max_tokens(BackendKind::AnthropicMessages, "claude-sonnet-4-6"),
-        )
-        .unwrap();
-
-        // System breakpoint still set.
-        assert_eq!(request["system"][0]["cache_control"]["type"], "ephemeral");
-        // No tools → no tools key.
-        assert!(request.get("tools").is_none());
-        // No messages → empty array, no crash.
-        assert_eq!(request["messages"].as_array().unwrap().len(), 0);
-    }
-
-    #[test]
-    fn summary_shaped_requests_preserve_all_systems_and_omit_tools() {
-        let messages = [
-            Message::System {
-                content: "primary".to_string(),
-            },
-            Message::System {
-                content: "agents".to_string(),
-            },
-            Message::User {
-                content: "historical checkpoint".to_string(),
-            },
-            Message::User {
-                content: "newly aged history".to_string(),
-            },
-            Message::User {
-                content: "compaction prompt".to_string(),
-            },
-        ];
-
-        let openai = openai_responses_request(
-            "model",
-            None,
-            &messages,
-            &[],
-            &test_thinking_levels(BackendKind::OpenAiResponses, "model"),
-        );
-        assert_eq!(openai["input"], serde_json::to_value(&messages).unwrap());
-        assert!(openai.get("tools").is_none());
-
-        for request in [
-            completions_chat_request(
-                "model",
-                None,
-                &messages,
-                &[],
-                &test_thinking_levels(BackendKind::FireworksChat, "model"),
-                &test_compat(BackendKind::FireworksChat, "model"),
-            ),
-            completions_chat_request(
-                "model",
-                None,
-                &messages,
-                &[],
-                &test_thinking_levels(BackendKind::TogetherChat, "model"),
-                &test_compat(BackendKind::TogetherChat, "model"),
-            ),
-            completions_chat_request(
-                "deepseek-v4-pro",
-                None,
-                &messages,
-                &[],
-                &test_thinking_levels(BackendKind::DeepSeekChat, "deepseek-v4-pro"),
-                &test_compat(BackendKind::DeepSeekChat, "deepseek-v4-pro"),
-            ),
-        ] {
-            assert_eq!(
-                request["messages"],
-                serde_json::to_value(&messages).unwrap()
-            );
-            assert!(request.get("tools").is_none());
-        }
-
-        let anthropic = anthropic_messages_request(
-            "claude-sonnet-4-6",
-            None,
-            &messages,
-            &[],
-            Some("1h"),
-            &test_thinking_levels(BackendKind::AnthropicMessages, "claude-sonnet-4-6"),
-            test_max_tokens(BackendKind::AnthropicMessages, "claude-sonnet-4-6"),
-        )
-        .unwrap();
-        assert_eq!(anthropic["system"][0]["text"], "primary\n\nagents");
-        assert_eq!(anthropic["messages"].as_array().unwrap().len(), 3);
-        assert_eq!(anthropic["messages"][0]["content"], "historical checkpoint");
-        assert_eq!(
-            anthropic["messages"][2]["content"][0]["text"],
-            "compaction prompt"
-        );
-        assert!(anthropic.get("tools").is_none());
-    }
-
-    #[test]
-    fn anthropic_response_tool_thinking_round_trips() {
-        let thinking = json!({
-            "type": "thinking",
-            "thinking": "",
-            "signature": "sig_1"
-        });
-        let redacted = json!({
-            "type": "redacted_thinking",
-            "data": "opaque"
-        });
-        let parsed = parse_anthropic_messages_response(
-            &json!({
-                "id": "msg_1",
-                "type": "message",
-                "role": "assistant",
-                "content": [
-                    thinking.clone(),
-                    redacted.clone(),
-                    {"type": "text", "text": "Need to inspect the file."},
-                    {
-                        "type": "tool_use",
-                        "id": "toolu_1",
-                        "name": "read",
-                        "input": {"path": "src/main.rs"}
-                    }
-                ],
-                "stop_reason": "tool_use",
-                "usage": {"input_tokens": 10, "output_tokens": 20}
-            }),
-            "https://api.anthropic.com/v1/messages",
-        )
-        .unwrap();
-
-        assert_eq!(
-            parsed.assistant.content.as_deref(),
-            Some("Need to inspect the file.")
-        );
-        assert_eq!(
-            parsed.assistant.reasoning_details,
-            Some(json!([thinking.clone(), redacted.clone()]))
-        );
-        assert_eq!(parsed.finish_reason.as_deref(), Some("tool_use"));
-        let tool_call = &parsed
-            .assistant
-            .tool_calls
-            .as_ref()
-            .expect("tool_use should become a tool call")[0];
-        assert_eq!(tool_call.id, "toolu_1");
-        assert_eq!(tool_call.function.name, "read");
-        assert_eq!(
-            serde_json::from_str::<Value>(&tool_call.function.arguments).unwrap(),
-            json!({"path": "src/main.rs"})
-        );
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 10);
-        assert_eq!(usage.output_tokens, 20);
-        assert_eq!(usage.cache_read_tokens, 0);
-        assert_eq!(usage.cache_write_tokens, 0);
-        assert_eq!(usage.orchestrator_context_tokens, 30);
-
-        let request = anthropic_messages_request(
-            "claude-opus-4-6",
-            None,
-            &[
-                Message::User {
-                    content: "please inspect".to_string(),
-                },
-                Message::Assistant {
-                    content: parsed.assistant.content.clone(),
-                    reasoning_text: None,
-                    reasoning_details: parsed.assistant.reasoning_details.clone(),
-                    tool_calls: parsed.assistant.tool_calls.clone(),
-                    model_origin: None,
-                    reasoning_field: None,
-                },
-                Message::Tool {
-                    tool_call_id: "toolu_1".to_string(),
-                    content: "file contents".to_string(),
-                },
-            ],
-            &[],
-            None,
-            &test_thinking_levels(BackendKind::AnthropicMessages, "claude-opus-4-6"),
-            test_max_tokens(BackendKind::AnthropicMessages, "claude-opus-4-6"),
-        )
-        .unwrap();
-
-        let assistant_blocks = request["messages"][1]["content"]
-            .as_array()
-            .expect("assistant content should be blocks");
-        assert_eq!(assistant_blocks[0], thinking);
-        assert_eq!(assistant_blocks[1], redacted);
-        assert_eq!(assistant_blocks[3]["type"], "tool_use");
-        assert_eq!(assistant_blocks[3]["input"], json!({"path": "src/main.rs"}));
-        assert_eq!(request["messages"][2]["role"], "user");
-        assert_eq!(request["messages"][2]["content"][0]["type"], "tool_result");
-        assert_eq!(
-            request["messages"][2]["content"][0]["tool_use_id"],
-            "toolu_1"
-        );
-    }
-
-    #[test]
-    fn deepseek_request_reasoning_is_driven_only_by_explicit_effort() {
-        let messages = [Message::Assistant {
-            content: Some("calling a tool".to_string()),
-            reasoning_text: Some("need current context".to_string()),
-            reasoning_details: None,
-            tool_calls: None,
-            model_origin: None,
-            reasoning_field: None,
-        }];
-        let levels = test_thinking_levels(BackendKind::DeepSeekChat, "deepseek-v4-pro");
-        let compat = test_compat(BackendKind::DeepSeekChat, "deepseek-v4-pro");
-        let absent =
-            completions_chat_request("deepseek-v4-pro", None, &messages, &[], &levels, &compat);
-        assert!(absent.get("thinking").is_none());
-        assert!(absent.get("reasoning_effort").is_none());
-        // DeepSeek's compat omits an explicit temperature.
-        assert!(absent.get("temperature").is_none());
-        assert_eq!(
-            absent["messages"][0]["reasoning_content"],
-            "need current context"
-        );
-
-        let disabled = completions_chat_request(
-            "deepseek-v4-pro",
-            Some(ReasoningEffort::None),
-            &messages,
-            &[],
-            &levels,
-            &compat,
-        );
-        assert_eq!(disabled["thinking"], json!({"type": "disabled"}));
-        assert!(disabled.get("reasoning_effort").is_none());
-
-        // The wire tier `max` for `xhigh` comes from the catalog map, not
-        // from adapter code (requests.rs carries no "max" literal).
-        for (effort, wire_effort) in [
-            (ReasoningEffort::High, "high"),
-            (ReasoningEffort::Xhigh, "max"),
-        ] {
-            let request = completions_chat_request(
-                "deepseek-v4-pro",
-                Some(effort),
-                &messages,
-                &[],
-                &levels,
-                &compat,
-            );
-            assert_eq!(request["thinking"], json!({"type": "enabled"}));
-            assert_eq!(request["reasoning_effort"], wire_effort);
-        }
-    }
-
-    #[test]
-    fn one_completions_builder_reproduces_every_provider_shape_from_compat() {
-        // S6 consolidation guard: a single builder, driven only by catalog
-        // compat data, reproduces the four distinct completions request
-        // shapes (DeepSeek, Fireworks, Together, Arcee) — including the
-        // per-provider reasoning replay field for unstamped (legacy)
-        // assistant messages.
-        let messages = [
-            Message::User {
-                content: "hi".to_string(),
-            },
-            Message::Assistant {
-                content: Some("prior".to_string()),
-                reasoning_text: Some("thought".to_string()),
-                reasoning_details: None,
-                tool_calls: None,
-                model_origin: None,
-                reasoning_field: None,
-            },
-        ];
-        let user = json!({"role": "user", "content": "hi"});
-
-        let deepseek = completions_chat_request(
-            "m",
-            Some(ReasoningEffort::High),
-            &messages,
-            &[],
-            &test_thinking_levels(BackendKind::DeepSeekChat, "m"),
-            &test_compat(BackendKind::DeepSeekChat, "m"),
-        );
-        assert_eq!(
-            deepseek,
-            json!({
-                "model": "m",
-                "messages": [
-                    user,
-                    {"role": "assistant", "content": "prior", "reasoning_content": "thought"}
-                ],
-                "thinking": {"type": "enabled"},
-                "reasoning_effort": "high"
-            }),
-            "DeepSeek: thinking dialect, no explicit temperature"
-        );
-
-        let fireworks = completions_chat_request(
-            "m",
-            Some(ReasoningEffort::High),
-            &messages,
-            &[],
-            &test_thinking_levels(BackendKind::FireworksChat, "m"),
-            &test_compat(BackendKind::FireworksChat, "m"),
-        );
-        assert_eq!(
-            fireworks,
-            json!({
-                "model": "m",
-                "messages": [
-                    user,
-                    {"role": "assistant", "content": "prior", "reasoning_content": "thought"}
-                ],
-                "temperature": 0.0,
-                "reasoning_effort": "high",
-                "reasoning_history": "preserved"
-            }),
-            "Fireworks: reasoning_effort + reasoning_history dialect"
-        );
-
-        let together = completions_chat_request(
-            "m",
-            Some(ReasoningEffort::High),
-            &messages,
-            &[],
-            &test_thinking_levels(BackendKind::TogetherChat, "m"),
-            &test_compat(BackendKind::TogetherChat, "m"),
-        );
-        assert_eq!(
-            together,
-            json!({
-                "model": "m",
-                "messages": [
-                    user,
-                    {"role": "assistant", "content": "prior", "reasoning": "thought"}
-                ],
-                "temperature": 0.0,
-                "reasoning": {"enabled": true},
-                "reasoning_effort": "high",
-                "chat_template_kwargs": {"clear_thinking": false}
-            }),
-            "Together: reasoning.enabled dialect, reasoning replay field"
-        );
-
-        // Arcee accepts no explicit effort levels (its map is empty), so the
-        // effort-free shape is the only reachable one.
-        let arcee = completions_chat_request(
-            "m",
-            None,
-            &messages,
-            &[],
-            &test_thinking_levels(BackendKind::ArceeApi, "m"),
-            &test_compat(BackendKind::ArceeApi, "m"),
-        );
-        assert_eq!(
-            arcee,
-            json!({
-                "model": "m",
-                "messages": [
-                    user,
-                    {"role": "assistant", "content": "prior", "reasoning_content": "thought"}
-                ],
-                "temperature": 0.0
-            }),
-            "Arcee: no thinking dialect"
-        );
-    }
-
-    #[test]
-    fn openai_compatible_request_schemas_honor_absent_none_and_supported_efforts() {
-        let messages = [Message::User {
-            content: "hi".into(),
-        }];
-
-        let fireworks_levels = test_thinking_levels(BackendKind::FireworksChat, "model");
-        let fireworks_compat = test_compat(BackendKind::FireworksChat, "model");
-        let fireworks_absent = completions_chat_request(
-            "model",
-            None,
-            &messages,
-            &[],
-            &fireworks_levels,
-            &fireworks_compat,
-        );
-        assert!(fireworks_absent.get("reasoning_effort").is_none());
-        assert!(fireworks_absent.get("reasoning_history").is_none());
-        assert!(fireworks_absent.get("tools").is_none());
-        assert_eq!(fireworks_absent["temperature"], json!(0.0));
-        let fireworks_none = completions_chat_request(
-            "model",
-            Some(ReasoningEffort::None),
-            &messages,
-            &[],
-            &fireworks_levels,
-            &fireworks_compat,
-        );
-        assert_eq!(fireworks_none["reasoning_effort"], "none");
-        assert_eq!(fireworks_none["reasoning_history"], "disabled");
-        for effort in [
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-        ] {
-            let request = completions_chat_request(
-                "model",
-                Some(effort),
-                &messages,
-                &[],
-                &fireworks_levels,
-                &fireworks_compat,
-            );
-            assert_eq!(request["reasoning_effort"], effort.as_str());
-            assert_eq!(request["reasoning_history"], "preserved");
-        }
-
-        let together_levels = test_thinking_levels(BackendKind::TogetherChat, "model");
-        let together_compat = test_compat(BackendKind::TogetherChat, "model");
-        let together_absent = completions_chat_request(
-            "model",
-            None,
-            &messages,
-            &[],
-            &together_levels,
-            &together_compat,
-        );
-        assert!(together_absent.get("reasoning").is_none());
-        assert!(together_absent.get("reasoning_effort").is_none());
-        assert!(together_absent.get("chat_template_kwargs").is_none());
-        assert!(together_absent.get("tools").is_none());
-        let together_none = completions_chat_request(
-            "model",
-            Some(ReasoningEffort::None),
-            &messages,
-            &[],
-            &together_levels,
-            &together_compat,
-        );
-        assert_eq!(together_none["reasoning"], json!({"enabled": false}));
-        assert!(together_none.get("reasoning_effort").is_none());
-        for effort in [
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-        ] {
-            let request = completions_chat_request(
-                "model",
-                Some(effort),
-                &messages,
-                &[],
-                &together_levels,
-                &together_compat,
-            );
-            assert_eq!(request["reasoning"], json!({"enabled": true}));
-            assert_eq!(request["reasoning_effort"], effort.as_str());
-            assert_eq!(
-                request["chat_template_kwargs"],
-                json!({"clear_thinking": false})
-            );
-        }
-
-        let openai_levels = test_thinking_levels(BackendKind::OpenAiResponses, "model");
-        let openai_absent = openai_responses_request("model", None, &messages, &[], &openai_levels);
-        assert!(openai_absent.get("reasoning").is_none());
-        assert!(openai_absent.get("tools").is_none());
-
-        let tools = [ToolDefinition {
-            def_type: "function".to_string(),
-            function: crate::types::FunctionDef {
-                name: "read".to_string(),
-                description: "Read a file".to_string(),
-                parameters: json!({"type": "object"}),
-            },
-        }];
-        assert!(
-            completions_chat_request(
-                "model",
-                None,
-                &messages,
-                &tools,
-                &fireworks_levels,
-                &fireworks_compat
-            )
-            .get("tools")
-            .is_some()
-        );
-        assert!(
-            completions_chat_request(
-                "model",
-                None,
-                &messages,
-                &tools,
-                &together_levels,
-                &together_compat
-            )
-            .get("tools")
-            .is_some()
-        );
-        assert!(openai_responses_request("model", None, &messages, &tools, &openai_levels)
-            .get("tools")
-            .is_some());
-        for effort in [
-            ReasoningEffort::None,
-            ReasoningEffort::Minimal,
-            ReasoningEffort::Low,
-            ReasoningEffort::Medium,
-            ReasoningEffort::High,
-            ReasoningEffort::Xhigh,
-        ] {
-            let request =
-                openai_responses_request("model", Some(effort), &messages, &[], &openai_levels);
-            assert_eq!(request["reasoning"]["effort"], effort.as_str());
-        }
-    }
-
-    #[test]
-    fn responses_input_items_expand_reasoning_and_tool_state() {
-        let items = responses_input_items(&[
-            Message::System {
-                content: "system".to_string(),
-            },
-            Message::Assistant {
-                content: Some("assistant text".to_string()),
-                reasoning_text: Some("hidden".to_string()),
-                reasoning_details: Some(json!([{
-                    "type": "reasoning",
-                    "id": "rs_1",
-                    "summary": [{"type": "summary_text", "text": "keep this"}]
-                }])),
-                tool_calls: Some(vec![ToolCall {
-                    id: "call_1".to_string(),
-                    call_type: "function".to_string(),
-                    function: FunctionCall {
-                        name: "read".to_string(),
-                        arguments: "{\"path\":\"src/main.rs\"}".to_string(),
-                    },
-                }]),
-                model_origin: None,
-                reasoning_field: None,
-            },
-            Message::Tool {
-                tool_call_id: "call_1".to_string(),
-                content: "tool output".to_string(),
-            },
-        ]);
-
-        assert_eq!(items.len(), 5);
-        assert_eq!(items[0]["role"], "system");
-        assert_eq!(items[1]["type"], "reasoning");
-        assert_eq!(items[2]["type"], "function_call");
-        assert_eq!(items[3]["role"], "assistant");
-        assert_eq!(items[4]["type"], "function_call_output");
-    }
-
-    #[test]
-    fn parses_deepseek_chat_output() {
-        let parsed = parse_completions_response(
-            &json!({
-                "choices": [
-                    {
-                        "finish_reason": "stop",
-                        "message": {
-                            "content": "done",
-                            "reasoning_content": "worked through it",
-                            "tool_calls": null
-                        }
-                    }
-                ],
-                "usage": {
-                    "prompt_tokens": 10,
-                    "completion_tokens": 20,
-                    "total_tokens": 30,
-                    "completion_tokens_details": {
-                        "reasoning_tokens": 9
-                    }
-                }
-            }),
-            "https://api.deepseek.com/chat/completions",
-            "reasoning_content",
-        )
-        .unwrap();
-
-        assert_eq!(parsed.assistant.content.as_deref(), Some("done"));
-        assert_eq!(
-            parsed.assistant.reasoning_text.as_deref(),
-            Some("worked through it")
-        );
-        assert_eq!(
-            parsed.assistant.reasoning_field.as_deref(),
-            Some("reasoning_content")
-        );
-        assert!(parsed.assistant.tool_calls.is_none());
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 10);
-        assert_eq!(usage.output_tokens, 20);
-        assert_eq!(usage.cache_read_tokens, 0);
-        assert_eq!(usage.cache_write_tokens, 0);
-        // S6: the unified parser reads the nested reasoning_tokens shape for
-        // every completions provider (previously DeepSeek/Fireworks/Arcee
-        // always parsed 0).
-        assert_eq!(usage.reasoning_tokens, 9);
-        assert_eq!(usage.orchestrator_context_tokens, 30);
-    }
-
-    #[test]
-    fn parses_openai_responses_output() {
-        let parsed = parse_openai_responses_response(
-            &json!({
-                "status": "completed",
-                "output": [
-                    {
-                        "type": "reasoning",
-                        "id": "rs_1",
-                        "summary": [{"type": "summary_text", "text": "thought summary"}]
-                    },
-                    {
-                        "type": "function_call",
-                        "call_id": "call_1",
-                        "name": "read",
-                        "arguments": "{\"path\":\"src/main.rs\"}"
-                    },
-                    {
-                        "type": "message",
-                        "content": [
-                            {"type": "output_text", "text": "hello world"}
-                        ]
-                    }
-                ],
-                "usage": {
-                    "input_tokens": 10,
-                    "output_tokens": 20,
-                    "total_tokens": 30,
-                    "output_tokens_details": {
-                        "reasoning_tokens": 7
-                    }
-                }
-            }),
-            "https://api.openai.com/v1/responses",
-        )
-        .unwrap();
-
-        assert_eq!(parsed.assistant.content.as_deref(), Some("hello world"));
-        assert_eq!(
-            parsed.assistant.reasoning_text.as_deref(),
-            Some("thought summary")
-        );
-        assert_eq!(
-            parsed
-                .assistant
-                .tool_calls
-                .as_ref()
-                .expect("tool calls should be parsed")
-                .len(),
-            1
-        );
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 10);
-        assert_eq!(usage.output_tokens, 20);
-        assert_eq!(usage.cache_read_tokens, 0);
-        assert_eq!(usage.cache_write_tokens, 0);
-        assert_eq!(usage.orchestrator_context_tokens, 30);
-    }
-
-    #[test]
-    fn parses_openai_responses_usage_with_cached_tokens() {
-        let parsed = parse_openai_responses_response(
-            &json!({
-                "status": "completed",
-                "output": [
-                    {"type": "message", "content": [{"type": "output_text", "text": "hi"}]}
-                ],
-                "usage": {
-                    "input_tokens": 100,
-                    "output_tokens": 50,
-                    "total_tokens": 150,
-                    "input_tokens_details": {"cached_tokens": 80},
-                    "output_tokens_details": {"reasoning_tokens": 10}
-                }
-            }),
-            "https://api.openai.com/v1/responses",
-        )
-        .unwrap();
-
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 20); // 100 - 80 cached
-        assert_eq!(usage.output_tokens, 50);
-        assert_eq!(usage.cache_read_tokens, 80);
-        assert_eq!(usage.cache_write_tokens, 0);
-        assert_eq!(usage.orchestrator_context_tokens, 150);
-
-        // Parsers never bill; the client attaches cost from catalog rates.
-        assert_eq!(usage.cost, TokenCostMicros::default());
-        // gpt-5.2 catalog rates ($/1M): 1.75 / 14 / 0.175 / 0.
-        let cost = calculate_cost(
-            &catalog::ModelCostRates {
-                input: 1.75,
-                output: 14.0,
-                cache_read: 0.175,
-                cache_write: 0.0,
-            },
-            None,
-            &usage,
-        );
-        assert_eq!(cost.input, 35);
-        assert_eq!(cost.output, 700);
-        assert_eq!(cost.cache_read, 14);
-        assert_eq!(cost.cache_write, 0);
-        assert_eq!(cost.total, 749);
-    }
-
-    #[test]
-    fn parses_anthropic_usage_with_cache_fields() {
-        let parsed = parse_anthropic_messages_response(
-            &json!({
-                "content": [{"type": "text", "text": "done"}],
-                "stop_reason": "end_turn",
-                "usage": {
-                    "input_tokens": 100,
-                    "output_tokens": 50,
-                    "cache_read_input_tokens": 200,
-                    "cache_creation_input_tokens": 30
-                }
-            }),
-            "https://api.anthropic.com/v1/messages",
-        )
-        .unwrap();
-
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 100);
-        assert_eq!(usage.output_tokens, 50);
-        assert_eq!(usage.cache_read_tokens, 200);
-        assert_eq!(usage.cache_write_tokens, 30);
-        assert_eq!(usage.orchestrator_context_tokens, 380); // 100 + 50 + 200 + 30
-
-        // Parsers never bill; the client attaches cost from catalog rates.
-        assert_eq!(usage.cost, TokenCostMicros::default());
-        // claude-opus-4-6 catalog rates ($/1M): 5 / 25 / 0.5 / 6.25.
-        let cost = calculate_cost(
-            &catalog::ModelCostRates {
-                input: 5.0,
-                output: 25.0,
-                cache_read: 0.5,
-                cache_write: 6.25,
-            },
-            None,
-            &usage,
-        );
-        assert_eq!(cost.input, 500);
-        assert_eq!(cost.output, 1_250);
-        assert_eq!(cost.cache_read, 100);
-        assert_eq!(cost.cache_write, 188, "187.5 micros rounds half-up");
-        assert_eq!(cost.total, 2_038);
-    }
-
-    #[test]
-    fn parses_chat_completions_usage_with_cached_tokens() {
-        let parsed = parse_completions_response(
-            &json!({
-                "choices": [{
-                    "finish_reason": "stop",
-                    "message": {"content": "done", "tool_calls": null}
-                }],
-                "usage": {
-                    "prompt_tokens": 100,
-                    "completion_tokens": 50,
-                    "total_tokens": 150,
-                    "prompt_tokens_details": {"cached_tokens": 60},
-                    "completion_tokens_details": {"reasoning_tokens": 5}
-                }
-            }),
-            "https://api.deepseek.com/chat/completions",
-            "reasoning_content",
-        )
-        .unwrap();
-
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 40); // 100 - 60 cached
-        assert_eq!(usage.output_tokens, 50);
-        assert_eq!(usage.cache_read_tokens, 60);
-        assert_eq!(usage.cache_write_tokens, 0);
-        assert_eq!(usage.orchestrator_context_tokens, 150);
-
-        // Parsers never bill; the client attaches cost from catalog rates.
-        assert_eq!(usage.cost, TokenCostMicros::default());
-        // deepseek-chat catalog rates ($/1M): 0.14 / 0.28 / 0.0028 / 0.
-        let cost = calculate_cost(
-            &catalog::ModelCostRates {
-                input: 0.14,
-                output: 0.28,
-                cache_read: 0.0028,
-                cache_write: 0.0,
-            },
-            None,
-            &usage,
-        );
-        assert_eq!(cost.input, 6, "5.6 micros rounds to 6");
-        assert_eq!(cost.output, 14);
-        assert_eq!(cost.cache_read, 0, "0.168 micros rounds to 0");
-        assert_eq!(cost.total, 20);
-    }
-
-    #[test]
-    fn response_without_usage_yields_none() {
-        let parsed = parse_openai_responses_response(
-            &json!({
-                "status": "completed",
-                "output": [
-                    {"type": "message", "content": [{"type": "output_text", "text": "hi"}]}
-                ]
-            }),
-            "https://api.openai.com/v1/responses",
-        )
-        .unwrap();
-
-        assert!(parsed.usage.is_none());
-    }
-
-    #[test]
-    fn parses_together_chat_response() {
-        let parsed = parse_completions_response(
-            &json!({
-                "choices": [
-                    {
-                        "finish_reason": "stop",
-                        "message": {
-                            "content": "The answer is 42.",
-                            "reasoning": "I need to think about this carefully...",
-                            "tool_calls": null
-                        }
-                    }
-                ],
-                "usage": {
-                    "prompt_tokens": 100,
-                    "completion_tokens": 50,
-                    "total_tokens": 150,
-                    "cached_tokens": 60,
-                    "reasoning_tokens": 25
-                }
-            }),
-            "https://api.together.ai/v1/chat/completions",
-            "reasoning",
-        )
-        .unwrap();
-
-        assert_eq!(
-            parsed.assistant.content.as_deref(),
-            Some("The answer is 42.")
-        );
-        assert_eq!(
-            parsed.assistant.reasoning_text.as_deref(),
-            Some("I need to think about this carefully...")
-        );
-        assert!(parsed.assistant.tool_calls.is_none());
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 40); // 100 - 60 cached
-        assert_eq!(usage.output_tokens, 50);
-        assert_eq!(usage.cache_read_tokens, 60);
-        assert_eq!(usage.cache_write_tokens, 0);
-        assert_eq!(usage.reasoning_tokens, 25);
-        assert_eq!(usage.orchestrator_context_tokens, 150);
-
-        // Parsers never bill; the client attaches cost from catalog rates.
-        assert_eq!(usage.cost, TokenCostMicros::default());
-        // MiniMax-M2.5 catalog rates ($/1M): 0.3 / 1.2 / 0.06 / 0.
-        let cost = calculate_cost(
-            &catalog::ModelCostRates {
-                input: 0.3,
-                output: 1.2,
-                cache_read: 0.06,
-                cache_write: 0.0,
-            },
-            None,
-            &usage,
-        );
-        assert_eq!(cost.input, 12);
-        assert_eq!(cost.output, 60);
-        assert_eq!(cost.cache_read, 4, "3.6 micros rounds to 4");
-        assert_eq!(cost.total, 76);
-    }
-
-    #[test]
-    fn parses_together_chat_response_nested_usage() {
-        let parsed = parse_completions_response(
-            &json!({
-                "choices": [{
-                    "finish_reason": "stop",
-                    "message": {
-                        "content": "The answer is 4.",
-                        "reasoning": "We need to calculate 2+2. That equals 4.",
-                        "tool_calls": null
-                    }
-                }],
-                "usage": {
-                    "prompt_tokens": 2618,
-                    "completion_tokens": 74,
-                    "total_tokens": 2692,
-                    "prompt_tokens_details": {"cached_tokens": 2560},
-                    "completion_tokens_details": {"reasoning_tokens": 71}
-                }
-            }),
-            "https://api.together.ai/v1/chat/completions",
-            "reasoning",
-        )
-        .unwrap();
-
-        assert_eq!(
-            parsed.assistant.content.as_deref(),
-            Some("The answer is 4.")
-        );
-        assert_eq!(
-            parsed.assistant.reasoning_text.as_deref(),
-            Some("We need to calculate 2+2. That equals 4.")
-        );
-        assert!(parsed.assistant.tool_calls.is_none());
-        let usage = parsed.usage.expect("usage should be parsed");
-        assert_eq!(usage.input_tokens, 58); // 2618 - 2560 cached
-        assert_eq!(usage.output_tokens, 74);
-        assert_eq!(usage.cache_read_tokens, 2560);
-        assert_eq!(usage.cache_write_tokens, 0);
-        assert_eq!(usage.reasoning_tokens, 71);
-        assert_eq!(usage.orchestrator_context_tokens, 2692);
-    }
-}
+mod tests;
